tutorial/requirements.txt
Django==1.8.6
Pygments==2.0.2
argparse==1.2.1
djangorestframework==3.3.1
wsgiref==0.1.2
